fix: correct error handling and value assignment in core modules
- Use std::io::Error::other for config serialization errors
- Replace filter_map with flatten for backlight device iterator
- Fix battery percentage calculation to avoid type casting issues

+++ unilii/plugins/Battery/src/lib.rs
@@ -70,7 +70,7 @@ impl Module for Battery {
                 }
             }
             ModuleUpdate::ProgressBar(value) => {
-                self.percentage = (value * 100.0) as f32;
+                self.percentage = value * 100.0;
             }
             ModuleUpdate::Icon(icon) => {
                 self.is_charging = icon == "charging";
